Add notmuch-config.el
I'm starting notmuch as a separate emacs window with:

/usr/bin/emacs -q --no-splash -l notmuch.el -l ~/.emacs.d/notmuch-config.el -f notmuch -mm -nbi

That way I can avoid any customizations made to the "coding" instance
of emacs interfering with the "email" instance.

At this point I'm simply copying the useful bits from init.el into
notmuch-config.el and adding notmuch-specific things to it.

Note that the above command line breaks Customize since there is no
init file to write to. I actually like this because it creates a kind
of read-only environment, which I think is important since this emacs
instance will be interacting with the outside world. That way, all
modifications need to be explicitly inserted into notmuch-config.el.