Hack for xref per-project stack
xref uses a single stack so when we switch to another project it mixes
the locations of the two.

This adds a hack to keep one stack per project and switch between
them.

If I remember correctly, there's a performance hit when switching
projects because I couldn't find a suitable extension point (hook) to
use.